reorganized admin folder into router, controller,validation files
